% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContrasts.R
\name{getContrasts}
\alias{getContrasts}
\title{Function to test cancer types.}
\usage{
getContrasts(data, test.method = "t.test")
}
\arguments{
\item{data}{}

\item{test.method}{the type of test. group1,group2: the compared groups in the pairwise tests. Available only when method = "t.test" or method = "wilcox.test".}
}
\value{
table of test results
}
\description{
Function to test cancer types.
}
\examples{
getContrasts(data = df_plot, test.method = "t.test")
}
