% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Read first sheet from an Excel file containing metadata}
\usage{
getData(input_file)
}
\arguments{
\item{input_file}{input an Excel file}
}
\value{
a data frame
}
\description{
Read first sheet from an Excel file containing metadata
}
\examples{
getData(input_file = "data/TCGA-CDR-SupplementalTableS1.xlsx")
}
