% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlot.R
\name{getPlot}
\alias{getPlot}
\title{Function to prepare data for visualization and plot a boxplot.}
\usage{
getPlot(data, cancer_type = "", patient_id = "")
}
\arguments{
\item{data}{}

\item{cancer_type}{(character vector): rows of interest (single or multiple; data from the ‘type’ column)}

\item{patient_id}{(character vector) - IDs of patients of interest (data from the ‘bcr_patient_barcode’ column)}
}
\value{
a boxplot and a data frame
}
\description{
Function to prepare data for visualization and plot a boxplot.
}
\examples{
getPlot(data = data, cancer_type = 'ACC|BRCA')
getPlot(data = data, patient_id = 'TTCGA-5N-A9KM|TCGA-BL-A0C8|TCGA-BL-A13I|TCGA-BL-A13J|TCGA-BL-A3JM|TCGA-BL-A5ZZ|TCGA-BT-A0S7|TCGA-BT-A0YX|TCGA-BT-A20J|TCGA-BT-A20N|TCGA-OR-A5J1|TCGA-OR-A5J2|TCGA-OR-A5J3|TCGA-OR-A5J4|TCGA-OR-A5J5|TCGA-OR-A5J6|TCGA-OR-A5J7|TCGA-OR-A5J8|TCGA-OR-A5J9|TCGA-OR-A5JA')
}
